requirements.txt
Flask
gunicorn
psycopg2-binary
pandas
python-dotenv
